requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.1.0
python-dotenv==1.0.1
cryptography==41.0.7

# Conector MySQL para SQLAlchemy (mysql+mysqlconnector)
mysql-connector-python==9.0.0

